CloudWatch/set-cw-loggroup-retention/requirements.txt
aws-cdk.core
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws-cdk.aws-iam